# docs/openapi/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"LegacyRestrictionDetailsResponse": {
				"properties": {
					"response": {
						"properties": {
							"errorMsg": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							},
							"respMsg": {
								"type": "string"
							},
							"respCodeMsg": {
								"type": "string"
							},
							"respCode": {
								"type": "string"
							},
							"respStatus": {
								"type": "boolean"
							}
						},
						"required": [
							"errorMsg",
							"errorCode",
							"respMsg",
							"respCodeMsg",
							"respCode",
							"respStatus"
						],
						"type": "object"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"response",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.boolean_": {
				"properties": {},
				"additionalProperties": {
					"type": "boolean"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"LegacyRestrictionDeniedResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"restrictions": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"result",
					"restrictions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyUserRestrictedRequest": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"customer",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFwcloudErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFwcloudRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyCustomerErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LegacyCustomerUpsertRequest": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"addr": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"web": {
						"type": "string"
					}
				},
				"required": [
					"customer",
					"name",
					"addr",
					"phone",
					"email",
					"web"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyCustomerResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"addr": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"web": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"created_by": {
						"type": "number",
						"format": "double"
					},
					"updated_by": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"addr",
					"phone",
					"email",
					"web",
					"created_at",
					"updated_at",
					"created_by",
					"updated_by"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyCustomerSummary": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyCustomerGetResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/LegacyCustomerResponse"
					},
					{
						"items": {
							"$ref": "#/components/schemas/LegacyCustomerSummary"
						},
						"type": "array"
					}
				]
			},
			"LegacyCustomerRequest": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"customer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyCustomerRestrictionResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"restrictions": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"result",
					"restrictions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyClusterGetRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					},
					"cluster": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud",
					"cluster"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyClusterRestrictionsResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"restrictions": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"result",
					"restrictions"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "FWCloud API",
		"version": "2.4.1",
		"description": "FWCloud RESTful API.",
		"contact": {}
	},
	"paths": {
...
		"/user/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"200": {
						"description": "Restriction details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyRestrictionDetailsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"response": {
												"respStatus": true,
												"respCode": "ACR_OK",
												"respCodeMsg": "Ok",
												"respMsg": "",
												"errorCode": "",
												"errorMsg": ""
											},
											"data": {}
										}
									}
								}
							}
						}
					},
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyRestrictionDeniedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check that there are no restrictions for user deletion.",
				"summary": "Restrictions for user deletion.",
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyUserRestrictedRequest"
							},
							"example": {
								"customer": 10,
								"user": 5
							}
						}
					}
				}
			}
		},
...
		"/fwcloud/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Check if the fwcloud indicated as a parameter has any deletion restriction.",
				"summary": "Check delete restrictions.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
...
		"/customer": {
			"post": {
				"operationId": "Store",
				"responses": {
					"204": {
						"description": "New customer"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1004,
											"msg": "Already exists with the same id"
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1005,
											"msg": "Already exists with the same name"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create new customer. Customers allow group users.",
				"summary": "New customer.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerUpsertRequest"
							},
							"example": {
								"customer": 1,
								"name": "FWCloud.net",
								"addr": "C/Carrasca, 7 - 03590 Altea (Alicante) - Spain",
								"phone": "+34 966 446 046",
								"email": "info@fwcloud.net",
								"web": "https://fwcloud.net"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "Update customer"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1005,
											"msg": "Already exists with the same name"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update customer's information.",
				"summary": "Update customer.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerUpsertRequest"
							},
							"example": {
								"customer": 2,
								"name": "FWCloud.net",
								"addr": "C/Carrasca, 7 - 03590 Altea (Alicante) - Spain",
								"phone": "+34 966 446 046",
								"email": "info@fwcloud.net",
								"web": "https://www.fwcloud.net"
							}
						}
					}
				}
			}
		},
		"/customer/get": {
			"put": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Get customer data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerGetResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 2,
											"name": "FWCloud.net",
											"addr": "C/Carrasca, 7 - 03590 Altea (Alicante) - Spain",
											"phone": "+34 966 446 046",
											"email": "info@fwcloud.net",
											"web": "https://fwcloud.net",
											"created_at": "2019-05-13T10:40:36.000Z",
											"updated_at": "2019-05-13T10:40:36.000Z",
											"created_by": 0,
											"updated_by": 0
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get customer data. If `customer` is empty, returns all customers with id and name.",
				"summary": "Get customer data.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerRequest"
							},
							"example": {
								"customer": 2
							}
						}
					}
				}
			}
		},
		"/customer/del": {
			"put": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "Delete customer"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerRestrictionResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete customer from the database.",
				"summary": "Delete customer.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerRequest"
							},
							"example": {
								"customer": 1
							}
						}
					}
				}
			}
		},
		"/customer/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerRestrictionResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check that there are no restrictions for customer deletion.",
				"summary": "Restrictions for customer deletion.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerRequest"
							},
							"example": {
								"customer": 10
							}
						}
					}
				}
			}
		},
...
		"/cluster/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterRestrictionsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"FirewallHasInterfaces": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check if there are restrictions for cluster deletion.",
				"summary": "Restrictions for cluster deletion.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 5
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}